catch up with latest oidf-box
Much of the OIDF stuff has moved into oidf-box, so we need lego changes
to catch up. In particular, this commit makes explicit use of
`acme.Identifier` to represent identifiers, rather than `string` as
before. lego woudl frequently parse the string identifiers and then
guess whether it was a DNS name, IP or OIDF entity, which works, but is
rather fragile.